+++ zcash_client_backend/src/fees/common.rs
@@ -7,7 +7,15 @@ use zcash_primitives::{
     },
 };
 
-use super::{sapling, ChangeError, ChangeValue, DustAction, DustOutputPolicy, TransactionBalance};
+use crate::ShieldedProtocol;
+
+use super::{
+    sapling as sapling_fees, ChangeError, ChangeValue, DustAction, DustOutputPolicy,
+    TransactionBalance,
+};
+
+#[cfg(feature = "orchard")]
+use super::orchard as orchard_fees;
 
 #[allow(clippy::too_many_arguments)]
 pub(crate) fn single_change_output_balance<
@@ -21,8 +29,8 @@ pub(crate) fn single_change_output_balance<
     target_height: BlockHeight,
     transparent_inputs: &[impl transparent::InputView],
     transparent_outputs: &[impl transparent::OutputView],
-    sapling_inputs: &[impl sapling::InputView<NoteRefT>],
-    sapling_outputs: &[impl sapling::OutputView],
+    sapling: &impl sapling_fees::BundleView<NoteRefT>,
+    #[cfg(feature = "orchard")] orchard: &impl orchard_fees::BundleView<NoteRefT>,
     dust_output_policy: &DustOutputPolicy,
     default_dust_threshold: NonNegativeAmount,
     change_memo: Option<MemoBytes>,
@@ -43,16 +51,67 @@ where
         .map(|t_out| t_out.value())
         .sum::<Option<_>>()
         .ok_or_else(overflow)?;
-    let sapling_in = sapling_inputs
+    let sapling_in = sapling
+        .inputs()
         .iter()
-        .map(|s_in| s_in.value())
+        .map(sapling_fees::InputView::<NoteRefT>::value)
         .sum::<Option<_>>()
         .ok_or_else(overflow)?;
-    let sapling_out = sapling_outputs
+    let sapling_out = sapling
+        .outputs()
         .iter()
-        .map(|s_out| s_out.value())
+        .map(sapling_fees::OutputView::value)
         .sum::<Option<_>>()
         .ok_or_else(overflow)?;
+
+    #[cfg(feature = "orchard")]
+    let orchard_in = orchard
+        .inputs()
+        .iter()
+        .map(orchard_fees::InputView::<NoteRefT>::value)
+        .sum::<Option<_>>()
+        .ok_or_else(overflow)?;
+    #[cfg(not(feature = "orchard"))]
+    let orchard_in = NonNegativeAmount::ZERO;
+
+    #[cfg(feature = "orchard")]
+    let orchard_out = orchard
+        .outputs()
+        .iter()
+        .map(orchard_fees::OutputView::value)
+        .sum::<Option<_>>()
+        .ok_or_else(overflow)?;
+    #[cfg(not(feature = "orchard"))]
+    let orchard_out = NonNegativeAmount::ZERO;
+
+    // TODO: implement a less naive strategy for selecting the pool to which change will be sent.
+    #[cfg(feature = "orchard")]
+    let (change_pool, sapling_change, orchard_change) =
+        if orchard_in > NonNegativeAmount::ZERO || orchard_out > NonNegativeAmount::ZERO {
+            // Send change to Orchard if we're spending any Orchard inputs or creating any Orchard outputs
+            (ShieldedProtocol::Orchard, 0, 1)
+        } else if sapling_in > NonNegativeAmount::ZERO || sapling_out > NonNegativeAmount::ZERO {
+            // Otherwise, send change to Sapling if we're spending any Sapling inputs or creating any
+            // Sapling outputs, so that we avoid pool-crossing.
+            (ShieldedProtocol::Sapling, 1, 0)
+        } else {
+            // For all other transactions, send change to Sapling.
+            // FIXME: Change this to Orchard once Orchard outputs are enabled.
+            (ShieldedProtocol::Sapling, 1, 0)
+        };
+    #[cfg(not(feature = "orchard"))]
+    let (change_pool, sapling_change) = (ShieldedProtocol::Sapling, 1);
+
+    #[cfg(feature = "orchard")]
+    let orchard_num_actions = orchard
+        .bundle_type()
+        .num_actions(
+            orchard.inputs().len(),
+            orchard.outputs().len() + orchard_change,
+        )
+        .map_err(ChangeError::BundleError)?;
+    #[cfg(not(feature = "orchard"))]
+    let orchard_num_actions = 0;
 
     let fee_amount = fee_rule
         .fee_required(
@@ -60,23 +119,20 @@ where
             target_height,
             transparent_inputs,
             transparent_outputs,
-            sapling_inputs.len(),
-            if sapling_inputs.is_empty() {
-                sapling_outputs.len() + 1
-            } else {
-                std::cmp::max(sapling_outputs.len() + 1, 2)
-            },
-            //Orchard is not yet supported in zcash_client_backend
-            0,
+            sapling.inputs().len(),
+            sapling
+                .bundle_type()
+                .num_outputs(
+                    sapling.inputs().len(),
+                    sapling.outputs().len() + sapling_change,
+                )
+                .map_err(ChangeError::BundleError)?,
+            orchard_num_actions,
         )
         .map_err(|fee_error| ChangeError::StrategyError(E::from(fee_error)))?;
 
-    let total_in = (t_in + sapling_in).ok_or_else(overflow)?;
-
-    let total_out = [t_out, sapling_out, fee_amount]
-        .iter()
-        .sum::<Option<NonNegativeAmount>>()
-        .ok_or_else(overflow)?;
+    let total_in = (t_in + sapling_in + orchard_in).ok_or_else(overflow)?;
+    let total_out = (t_out + sapling_out + orchard_out + fee_amount).ok_or_else(overflow)?;
 
     let proposed_change = (total_in - total_out).ok_or(ChangeError::InsufficientFunds {
         available: total_in,
@@ -101,7 +157,7 @@ where
                     })
                 }
                 DustAction::AllowDustChange => TransactionBalance::new(
-                    vec![ChangeValue::sapling(proposed_change, change_memo)],
+                    vec![ChangeValue::new(change_pool, proposed_change, change_memo)],
                     fee_amount,
                 )
                 .map_err(|_| overflow()),
@@ -113,7 +169,7 @@ where
             }
         } else {
             TransactionBalance::new(
-                vec![ChangeValue::sapling(proposed_change, change_memo)],
+                vec![ChangeValue::new(change_pool, proposed_change, change_memo)],
                 fee_amount,
             )
             .map_err(|_| overflow())